+++ apps/jido_murmur/test/jido_murmur/runner_test.exs
@@ -2,6 +2,7 @@ defmodule JidoMurmur.RunnerTest do
   use JidoMurmur.Case, async: false
 
   alias JidoMurmur.Ingress
+  alias JidoMurmur.Observability.Store
   alias JidoMurmur.Runner
 
   setup do
@@ -82,9 +83,7 @@ defmodule JidoMurmur.RunnerTest do
       :ets.new(:jido_murmur_active_runners, [:set, :public, :named_table])
     end
 
-    unless :ets.whereis(:jido_murmur_obs_conversations) != :undefined do
-      :ets.new(:jido_murmur_obs_conversations, [:named_table, :public, :set])
-    end
+    Store.create_tables()
   rescue
     ArgumentError -> :ok
   end