+++ apps/jido_murmur/test/jido_murmur/table_owner_test.exs
@@ -9,7 +9,6 @@ defmodule JidoMurmur.TableOwnerTest do
         # Clean up any stale tables
         for table <- [
               :jido_murmur_active_runners,
-              :jido_murmur_obs_conversations,
               :jido_murmur_obs_turns,
               :jido_murmur_obs_agent_turns,
               :jido_murmur_obs_llm_spans,
@@ -34,7 +33,6 @@ defmodule JidoMurmur.TableOwnerTest do
   describe "start_link/1" do
     test "creates ETS tables on init" do
       assert :ets.whereis(:jido_murmur_active_runners) != :undefined
-      assert :ets.whereis(:jido_murmur_obs_conversations) != :undefined
       assert :ets.whereis(:jido_murmur_obs_turns) != :undefined
       assert :ets.whereis(:jido_murmur_obs_llm_spans) != :undefined
       assert :ets.whereis(:jido_murmur_obs_sessions) != :undefined
@@ -54,14 +52,12 @@ defmodule JidoMurmur.TableOwnerTest do
 
     test "tables are public" do
       assert :ets.info(:jido_murmur_active_runners)[:protection] == :public
-      assert :ets.info(:jido_murmur_obs_conversations)[:protection] == :public
       assert :ets.info(:jido_murmur_obs_turns)[:protection] == :public
       assert :ets.info(:jido_murmur_obs_sessions)[:protection] == :public
     end
 
     test "tables are named tables" do
       assert :ets.info(:jido_murmur_active_runners)[:named_table] == true
-      assert :ets.info(:jido_murmur_obs_conversations)[:named_table] == true
       assert :ets.info(:jido_murmur_obs_turns)[:named_table] == true
       assert :ets.info(:jido_murmur_obs_sessions)[:named_table] == true
     end